+++ example/lib/config/env.dart
@@ -1,0 +1,3 @@
+class ENV {
+  static const TRACKER_HOST = 'http://192.168.11.7:9000';
+}