+++ ios/Classes/streams/AuthorizationStreamHandler.h
@@ -1,0 +1,11 @@
+//
+//  AuthenticationStreamHandler.h
+//  Pods
+//
+//  Created by Christopher Scott on 2019-11-27.
+//
+#import "StreamHandler.h"
+
+@interface AuthorizationStreamHandler : StreamHandler
+@end
+