+++ lib/models/config.dart
@@ -233,7 +233,7 @@ class Config {
 
   /// __`[Android only]`__ Enable high-accuracy for **geofence-only** mode (See [BackgroundGeolocation.startGeofences]).
   ///
-  /// Defaults to `false`.  Runs Android's [BackgroundGeolocation.startGeofences] with a///foreground service* (along with its corresponding persitent notification;  See [Notification] for a list of available notification config options, including [notification.text], [notification.title]).
+  /// Defaults to `false`.  Runs Android's [BackgroundGeolocation.startGeofences] with a///foreground service* (along with its corresponding persitent notification;  See [Notification] for a list of available notification config options, including [Notification.text], [Notification.title]).
   ///
   /// Configuring `geofenceModeHighAccuracy: true` will make Android geofence triggering///*far more responsive**.  In this mode, the usual config options to control location-services will be applied:
   ///
@@ -410,6 +410,13 @@ class Config {
   /// __See also:__ __HTTP Guide__ at [HttpEvent].
   ///
   int persistMode;
+
+  /// Disable [autoSync] HTTP requests when device is connected to a Cellular connection.
+  /// Defaults to `false`.  Set `true` to allow [autoSync] only when device is connected to Wifi.
+  ///
+  /// __WARNING__ This option is ignored when manually invoking [BackgroundGeolocation.sync].
+  ///
+  bool disableAutoSyncOnCellular;
 
   /// The HTTP method to use when creating an HTTP request to your configured [url].
   ///
@@ -854,6 +861,90 @@ class Config {
   /// ```
   ///
   int httpTimeout;
+
+  /// Encrypt location data in the SDK's SQLite datbase and HTTP requests (__`AES-256-CBC`__).
+  ///
+  /// Defaults to `false`.  When enabled, the SDK will encrypt location data in its SQLite database.  When executing HTTP requests, the SDK will encrypt the entire request payload and encode the result as `Base64`.
+  ///
+  /// ```dart
+  /// BackgroundGeolocation.ready(Config(
+  ///   encrypt: true
+  /// ));
+  /// ```
+  ///
+  /// ## Encryption Password
+  ///
+  /// The SDK's encryption stack requires a configurable encryption *password*.
+  ///
+  /// ### iOS
+  ///
+  /// In your __`Info.plist`__, Add the `String` key `BACKGROUND_GEOLOCATION_ENCRYPTION_PASSWORD`.
+  ///
+  /// ![](https://www.dropbox.com/s/amea0siu9mxroh3/ios-encryption_password.png?dl=1)
+  ///
+  /// ### Android
+  ///
+  /// In your __`AndroidManifest.xml`__, add the following `<meta-data>` element:
+  ///
+  /// ```xml
+  /// <application>
+  ///     .
+  ///     .
+  ///     .
+  ///     <meta-data
+  ///       android:name="com.transistorsoft.locationmanager.ENCRYPTION_PASSWORD"
+  ///       android:value="your secret encryption password"
+  ///     />
+  /// </application>
+  /// ```
+  ///
+  /// ## RNCryptor Encryption Stack
+  ///
+  /// The SDK uses the [RNCryptor Encryption Stack](https://github.com/RNCryptor/RNCryptor-Spec/blob/master/RNCryptor-Spec-v3.md).  See [RNCypto](https://github.com/RNCryptor) for a list of available language implementations.
+  ///
+  /// After decoding the `Base64`-encoded data from the HTTP request body, you'll have a binary payload.  Extract bytes as follows:
+  ///
+  /// ![](https://dl.dropbox.com/s/owp61pt3cqfij16/RNCrypto-DataFormat-Spec.png?dl=1)
+  ///
+  /// | Name             | Description                                       |
+  /// |------------------|---------------------------------------------------|
+  /// | `version`        | (1 byte): Data format version. (Currently `3`).   |
+  /// | `options`        | (1 byte): bit 0 - uses password (Always `1`).     |
+  /// | `encryptionSalt` | (8 bytes)                                         |
+  /// | `HMACSalt`       | (8 bytes)                                         |
+  /// | `IV`             | (16 bytes)                                        |
+  /// | `ciphertext`     | (variable) -- Encrypted in CBC mode               |
+  /// | `HMAC`           | (32 bytes)
+  ///
+  /// See [here](https://gist.github.com/christocracy/f814dd35cfd9eced5d4de3025c38333c) for a NodeJS-based decryption example.
+  ///
+  /// ### Password-based decryption (abstract language)
+  ///
+  /// ```
+  /// def Decrypt(Password, Message) =
+  ///   (Version,Options,EncryptionSalt,HMACSalt,IV,Ciphertext,HMAC) = Split(Message)
+  ///     EncryptionKey = PKBDF2(EncryptionSalt, 32 length, 10k iterations, Password)
+  ///     HMACKey = PKBDF2(HMACSalt, 32 length, 10k iterations, password)
+  ///     Header = 3 || 1 || EncryptionSalt || HMACSalt || IV
+  ///     Plaintext = AES256Decrypt(Ciphertext, ModeCBC, IV, EncryptionKey)
+  ///     ComputedHMAC = HMAC(Header || Ciphertext, HMACKey, SHA-256)
+  ///     if ConsistentTimeEqual(ComputedHMAC, HMAC) return Plaintext else return Error
+  /// ```
+  ///
+  /// 1. Pull apart the pieces as described in the data format.
+  /// 1. Generate the encryption key using PBKDF2 (see your language docs for how to call this). Pass the password as a string, the random encryption salt, 10,000 iterations, and SHA-1 PRF. Request a length of 32 bytes.
+  /// 1. Generate the HMAC key using PBKDF2 (see your language docs for how to call this). Pass the password as a string, the random HMAC salt, 10,000 iterations, and SHA-1 PRF. Request a length of 32 bytes.
+  /// 1. Decrypt the data using the encryption key (above), the given IV, AES-256, and the CBC mode. This is the default mode for almost all AES encryption libraries.
+  /// 1. Pass your header and ciphertext to an HMAC function, along with the HMAC key (above), and the PRF "SHA-256" (see your library's docs for what the names of the PRF functions are; this might also be called "SHA-2, 256-bits").
+  /// 1. Compare the computed HMAC with the expected HMAC using a constant time equality function (see below). If they are equal, return the plaintext. Otherwise, return an error
+  ///
+  /// Note: The RNCryptor format v3 uses SHA-1 for PBKDF2, but SHA-256 for HMAC.
+  ///
+  bool encrypt;
+
+  /// Authorization
+  ///
+  Authorization authorization;
 
   /// Controls whether to continue location-tracking after application is **terminated**.
   ///
@@ -1777,6 +1868,48 @@ class Config {
   /// __@deprecated.__  Use [Notification.channelName]
   String notificationChannelName;
 
+  /// *Convenience* option to automatically configures the SDK to upload locations to the Transistor Software demo server at http://tracker.transistorsoft.com (or your own local instance of [background-geolocation-console](https://github.com/transistorsoft/background-geolocation-console))
+  ///
+  /// See [TransistorAuthorizationToken].  This option will **automatically configures** the [url] to point at the Demo server as well as well as the required [Authorization] configuration.
+  ///
+  ///
+  /// ## Example
+  /// ```dart
+  /// TransistorAuthorizationToken token = await
+  ///   TransistorAuthorizationToken.findOrCreate('my-company-name', 'my-username');
+  ///
+  /// BackgroundGeolocation.ready(Config(
+  ///   transistorAuthorizationToken: token
+  /// ));
+  /// ```
+  ///
+  /// This *convenience* option merely performs the following [Authorization] configuration *automatically* for you:
+  ///
+  /// ```dart
+  /// // Base url to Transistor Demo Server.
+  /// const String url = 'http://tracker.transistorsoft.com';
+  ///
+  /// // Register for an authorization token from server.
+  /// TransistorAuthorizationToken token = await
+  ///   TransistorAuthorizationToken.findOrCreate('my-company-name', 'my-username');
+  ///
+  /// BackgroundGeolocation.ready(Config(
+  ///   url: "$url/v2/locations",
+  ///   authorization: Authorization(
+  ///     strategy: "JWT",
+  ///     accessToken: token.accessToken,
+  ///     refreshToken: token.refreshToken,
+  ///     refreshUrl: "$url/v2/refresh_token",
+  ///     refreshPayload: {
+  ///       "refresh_token": "{refreshToken}"
+  ///     },
+  ///     expires: token.expires
+  ///   )
+  /// ));
+  /// ```
+  ///
+  TransistorAuthorizationToken transistorAuthorizationToken;
+
   Config(
       {
       // Geolocation Options
@@ -1808,6 +1941,7 @@ class Config {
       this.headers,
       this.extras,
       this.autoSync,
+      this.disableAutoSyncOnCellular,
       this.autoSyncThreshold,
       this.batchSync,
       this.maxBatchSize,
@@ -1817,6 +1951,8 @@ class Config {
       this.maxRecordsToPersist,
       this.locationsOrderDirection,
       this.httpTimeout,
+      this.encrypt,
+      this.authorization,
       // Application
       this.stopOnTerminate,
       this.startOnBoot,
@@ -1874,7 +2010,8 @@ class Config {
       this.notificationColor,
       this.notificationSmallIcon,
       this.notificationLargeIcon,
-      this.notificationChannelName});
+      this.notificationChannelName,
+      this.transistorAuthorizationToken});
 
   Config set(String key, dynamic value) {
     if (_map == null) {
@@ -1890,6 +2027,13 @@ class Config {
     }
 
     Map config = {};
+
+    // Were we provided a Transistor token?  Auto-config the url and authorization.
+    if (transistorAuthorizationToken != null) {
+      url = transistorAuthorizationToken.locationsUrl;
+      authorization = transistorAuthorizationToken.authorizationConfig;
+    }
+
     // Geolocation Options
     if (desiredAccuracy != null) config['desiredAccuracy'] = desiredAccuracy;
     if (distanceFilter != null) config['distanceFilter'] = distanceFilter;
@@ -1929,6 +2073,8 @@ class Config {
     if (headers != null) config['headers'] = headers;
     if (extras != null) config['extras'] = extras;
     if (autoSync != null) config['autoSync'] = autoSync;
+    if (disableAutoSyncOnCellular != null)
+      config['disableAutoSyncOnCellular'] = disableAutoSyncOnCellular;
     if (autoSyncThreshold != null)
       config['autoSyncThreshold'] = autoSyncThreshold;
     if (batchSync != null) config['batchSync'] = batchSync;
@@ -1941,6 +2087,8 @@ class Config {
     if (locationsOrderDirection != null)
       config['locationsOrderDirection'] = locationsOrderDirection;
     if (httpTimeout != null) config['httpTimeout'] = httpTimeout;
+    if (encrypt != null) config['encrypt'] = encrypt;
+    if (authorization != null) config['authorization'] = authorization.toMap();
     // Application
     if (stopOnTerminate != null) config['stopOnTerminate'] = stopOnTerminate;
     if (startOnBoot != null) config['startOnBoot'] = startOnBoot;
@@ -2050,67 +2198,20 @@ class Config {
           priority: notificationPriority);
       config['notification'] = notification.toMap();
     }
+
     return config;
   }
 
-  static final DeviceInfoPlugin deviceInfo = new DeviceInfoPlugin();
-
-  /// Returns a Map suitable for attaching to [params] and posting to the Transistor Software Demo Server at `http://tracker.transistorsoft.com`.
-  ///
-  ///
-  /// ## Example
-  ///
-  /// ```dart
-  /// // Required by Demo Server.
-  /// Map deviceParams = await bg.Config.deviceParams;
-  /// String username = 'my-unique-username'; // Eg Github username
-  ///
-  /// BackgroundGeolocation.ready(Config(
-  ///   url: 'http://tracker.transistorsoft.com/locations/$username',
-  ///   params: deviceParams
-  /// ));
-  /// ```
-  ///
-  /// View your tracking at [http://tracker.transistorsoft.com/my-unique-username](http://tracker.transistorsoft.com/my-unique-username).
-  ///
-  /// ![](https://dl.dropboxusercontent.com/s/3abuyyhioyypk8c/screenshot-tracker-transistorsoft.png?dl=1)
+  /// __@deprecated__: Use [transistorAuthorizationToken]
   ///
   static Future<Map<String, dynamic>> get deviceParams async {
-    Map<String, dynamic> response;
-    try {
-      RegExp re = new RegExp(r"[\s\.,]");
-      if (defaultTargetPlatform == TargetPlatform.android) {
-        AndroidDeviceInfo info = await deviceInfo.androidInfo;
-        String uuid =
-            '${info.model}-${info.version.release}'.replaceAll(re, '-');
-        response = {
-          'device': {
-            'uuid': uuid,
-            'model': info.model,
-            'platform': 'Android',
-            'manufacturer': info.manufacturer,
-            'version': info.version.release,
-            'framework': 'Flutter'
-          }
-        };
-      } else if (defaultTargetPlatform == TargetPlatform.iOS) {
-        IosDeviceInfo info = await deviceInfo.iosInfo;
-        String uuid =
-            '${info.utsname.machine}-${info.systemVersion}'.replaceAll(re, '-');
-        response = {
-          'device': {
-            'uuid': uuid,
-            'model': info.utsname.machine,
-            'platform': 'iOS',
-            'manufacturer': 'Apple',
-            'version': info.systemVersion,
-            'framework': 'Flutter'
-          }
-        };
-      }
-    } on PlatformException {
-      response = {'Error:': 'Failed to get platform version.'};
-    }
-    return response;
+    DeviceInfo deviceInfo = await DeviceInfo.getInstance();
+    Map map = deviceInfo.toMap();
+    RegExp re = new RegExp(r"[\s\.,]");
+    String uuid =
+        '${deviceInfo.model}-${deviceInfo.version}'.replaceAll(re, '-');
+    map['uuid'] = uuid;
+    map['framework'] = 'flutter';
+    return map;
   }
 }